requirements.txt
Django>=4.2
djangorestframework
drf-yasg          
psycopg2-binary   
pypdf             
openai            
python-dotenv    
tiktoken         
numpy
requests